add cms files

+++ application/controllers/admin/article.php
@@ -1,27 +1,30 @@
 <?php
 
-class Article extends Admin_Controller {
+class Article extends CI_Controller {
 
 
 	public function __construct() {
 		parent::__construct();
+		$this->data['meta_title'] = 'Dashboard CMS';
+		$this->load->helper('form');
+		$this->load->library('form_validation');
+		$this->load->library('session');
 		$this->load->model('article_m');
 	}
 
 	
 	public function index() {
 		//Get all articles
-		$this->data['articles'] = $this->article_m->get();
+		$this->data['articles'] = $this->article_m->get_news();
 		//Load view
 		$this->data['subview'] = 'admin/article/index';
 		$this->load->view('admin/_layout_main', $this->data);
 	}
 
 	public function edit($id = null) {
-
 		//Get a article or set a new one
 		if ($id) {
-			$this->data['article'] = $this->article_m->get($id);
+			$this->data['article'] = $this->article_m->get_news_by_id($id);
 			count($this->data['article']) || $this->data['errors'][] 
 			= 'article could not be found';
 		} else {
@@ -34,15 +37,21 @@ class Article extends Admin_Controller {
 		//Process the form
 		$this->form_validation->set_rules($rules);
 		//save article
+		foreach ($_POST as $key => $value) {
+			if(in_array($key,array('pubdate','title','tag','body'))) {
+				$data[$key] = $value;	
+			}
+			
+		}
+
 		if ($this->form_validation->run() == true) {
-			$data = $this->article_m->array_from_post(array(
-				'title',
-				'tag',
-				'body',
-				'pubdate'
-			));
-			$this->article_m->save($data,$id);
-			redirect('admin/article');
+			if($id) {
+				$this->article_m->update($id,$data);
+				redirect('admin/article');
+			} else {
+				$this->article_m->create_new($data);
+				redirect('admin/article');
+			}
 		}
 
 		//Load the view
@@ -51,6 +60,11 @@ class Article extends Admin_Controller {
 	}
 
 	public function delete($id) {
+		
+		if(!$id) {
+			return false;
+		}
+		
 		$this->article_m->delete($id);
 		redirect('admin/article');
 	}
